+++ app/__init__.py
@@ -2,4 +2,4 @@
 # Copyright (C) 2025-2026 TG11
 """Vulpfin Flowboard application package."""
 
-__version__ = "2.4.1"
+__version__ = "2.5.0"